session timeout field added

+++ sdk/src/main/java/com/exponea/sdk/models/ExponeaConfiguration.kt
@@ -16,7 +16,7 @@ data class ExponeaConfiguration(
         // Maximum retries value to flush data to api.
         var maxTries: Int = 10,
         // Timeout session value considered for app usage.
-        var sessionTimeout: Int = 20,
+        var sessionTimeout: Double = 20.0,
         // Flag to control automatic tracking for In-App purchases
         var automaticSessionTracking: Boolean = true,
         // Flag to control if the App will handle push notifications automatically.
